langchain-python-practice/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
statsmodels
python-dotenv
pymongo
ipykernel
langchain
openai
langchain_community
python-dotenv
langchain-openai
langchain_google_genai
google-cloud-aiplatform
vertexai
google.generativeai
